efi/Makefile: Use original efi linker script from gnu-efi lib
There is a issue with the efi/elf_aarch64_efi.lds script. It caused the
fwupaa64.efi corrupted and can be not loaded by UEFI.
The linker script should be used directly from gnu-efi library instead of
copying.

To compile this fwupdate, the gnu-efi is required. For CentOS 8,
installing gnu-efi with the following command:
$ dnf --enablerepo=PowerTools install gnu-efi*

Signed-off-by: Nhi Pham <npham@amperecomputing.com>